BF: do not bother workarounding enum for schemaKey if there is None
Somehow I do not run into this behavior with pydantic 1.8.2 when using custom
metaclass.  Might be a sign of trouble but all tests pass, so I decided just
to condition on having "enum".   But then saw that resultant schema lacks "const"
so reworked to add "const" even if not enum.  May be that is not desired?