RF: make archive URL to be "opt-in" and absent by default
It will be taken from the environment variable DJANGO_DANDI_WEB_APP_URL
if specified, as it is now set in the teraform configuration:

    $> git grep DJANGO_DANDI_WEB_APP_URL
    terraform/api.tf:    DJANGO_DANDI_WEB_APP_URL                     = "https://dandiarchive.org"
    terraform/staging_pipeline.tf:    DJANGO_DANDI_WEB_APP_URL                     = "https://gui-staging.dandiarchive.org"

so with this change we would get deployed setups ensuring that metadata
does have corresponding to that deployment metadata.

Note that  Resource.repository  is also set based on the
DJANGO_DANDI_WEB_APP_URL environment variable.
It is a datatype for CommonModel.relatedResource . To be generic, I do not
think we even should impose any specific default value there.