Progress DANDI_SCHEMA_VERSION to 0.6.3
I was wondering if it should become 0.7.0 but since this change should not
really be breaking much really, I decided to just keep it within 0.6.x series
without even bothering about any migration.  Any changes should be done in
dandi-cli or dandi-archive (yet to be decided) to fill in the repository.

Note that it is marked readOnly, but I would assume it is read only for the
meditor UI and tooling can still change it in the record