Introduce ContentType as Blob vs Zarr and rename blobDateModified to contentDateModified
This is just an initial attempt open for discussion.

I ran into "blobDateModified" in a zarr metadata and it raised my eyebrow since
that is not really appropriate and confusing. Hence I decided to look into
generalization.  I also thought that it would be valuable to make "type"
of the content Asset points to explicit, although that could lead to
inconsistencies since information is somewhat redundant with encodingFormat and
potentially could also be deduced from contenUrl since we have different end
points on S3, etc.

Nevertheless I think it might be better to make it explicit. Or at least we
have to rename blobDateModified.

- ContenType name is quite suboptimal since there is a standard HTTP header
  https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type
  and thus we could potential confusion.

  But we should keep it a "Type" (not e.g. a Class) to be consistent with
  other type definitions among models.

  So the other part we could try to vary is "Content". Possible alternatives are
  "Object", "Data", "Resource"

- ATM we call all Zarrs just Zarr but it is a "ZarrFolder" really.  I wonder if
  it would be time to start to introduce differentiation here by making it
  "ZarrFolder", as later we might get "ZarrHDF5" or alike