fix: remove the "DOI" from alternateIdentifiers
It kinda historically dragged through time from the original
1056afb49fd945afc471e200d782bbff9de43cf1 in dandi-cli where it
was added to "identifiers".  As the wise @tmorell has mentioned,
since it is the datacite which is to provide DOI, it would ignore
DOI "alternateIdentifiers".  I think it makes sensse overall, although
there could potentially be multiple DOIs for a single dandiset -- nothing
in DOI principle forbids it. But since original purpose here is not
clear -- we better just strip it away since it should be the DOI
minted by datacite fabrica as the one for the PublishedDandiset