Use `title` in place of `examples`
The `title` property is rendered by VJSF as dimmed placeholder text
inside the field itself, which is a subtle but clear hint to the user to
format the name correctly. This is simpler and less obtrusive than using
`examples`, which causes VJSF to instead render a dropdown menu
containing the examples. This prevents editing of an existing name and
implies that the example names are the only suggested names in the
system.