Duplicate Publishable attributes in CommonModel + DOI in Dandiset as Optional
That is necessary to later be able to make our model more restrictive
and disallow addition of arbitrary attributes as it does now.

ATM any Draft version of a dandiset which was already published at least once,
would have those attributes although base Dandiset (similar in Asset model;
@candleindark would check) does not have them defined. This is what this PR
fixes.

More details on how those were found and in which dandisets could be obtained from
https://github.com/dandi/dandi-schema-status/blob/main/reports/diff_reports/dandiset/jsonschema_errs_summary.md#errs-diff-detailed-tables

This should be "backward compatible" change at the level of the schema,
as any prior valid record would remain valid. Hence just boosting the "patch"
portion of the version