feat: define `PublishedDandiset.doi` on condition
If `DANDI_DOI_PATTERN` is not None define
`PublishedDandiset.doi` with its previous definition.
Otherwise, restrict `PublishedDandiset.doi` to be an
empty string and default it to the empty string.
With this definition, `PublishedDandiset.doi` always
exits regardless of the value of `DANDI_DOI_PATTERN`.
Yet, when `DANDI_DOI_PATTERN` is `None`,
`PublishedDandiset.doi` is always the empty string
indicating that the DOI doesn't exist.